prepare for the new expander
move some function and delete the expander
