+++ sources/environment2.c
@@ -6,20 +6,44 @@
 /*   By: tchoquet <tchoquet@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/19 17:25:41 by tchoquet          #+#    #+#             */
-/*   Updated: 2023/08/21 15:20:25 by tchoquet         ###   ########.fr       */
+/*   Updated: 2023/08/24 14:30:05 by tchoquet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "environment.h"
 
-t_bool		is_env_key_equal(void *entry, void *key);
-void		delete_entry(void *v_entry);
+char		*set_error_return(int *error, int value);
+t_env_entry	str_to_env_entry(const char *str);
 
 t_env_list	*lstenv_chr(const char *key, t_bool create);
 
-int			analize_keyval(const char *keyval, t_env_entry *result);
-t_bool		is_valid_env_key(const char *str);
-char		*set_error_return(int *error, int value);
+static int	analize_keyval(const char *keyval, t_env_entry *input);
+
+int	set_env_single_str(const char *keyval, t_bool export)
+{
+	t_env_entry	input;
+	t_env_list	*founded;
+	int			temp_ret;
+
+	temp_ret = is_valid_keyval(keyval);
+	if (temp_ret != 0)
+		return (temp_ret);
+	input = str_to_env_entry(keyval);
+	if (input.key == NULL)
+		return (MALLOC_ERROR);
+	founded = lstenv_chr(input.key, true);
+	free(input.key);
+	if (founded == NULL)
+	{
+		free(input.value);
+		return (MALLOC_ERROR);
+	}
+	free(founded->data->value);
+	founded->data->value = input.value;
+	if (export == true)
+		founded->data->is_export = true;
+	return (0);
+}
 
 char	*get_env_create(const char *keyval, int *error_code)
 {
@@ -50,13 +74,15 @@ char	*get_env_create(const char *keyval, int *error_code)
 	return (set_error_return(error_code, MALLOC_ERROR));
 }
 
-int	delete_env(const char *key)
+static int	analize_keyval(const char *keyval, t_env_entry *input)
 {
-	if (is_valid_env_key(key) == false)
-		return (BAD_ENVIRONMENT_KEY);
-	lst_delif(
-		(t_list **)(get_lstenv()),
-		&delete_entry, &is_env_key_equal,
-		(void *)key);
+	int	temp_ret;
+
+	temp_ret = is_valid_keyval(keyval);
+	if (temp_ret != 0)
+		return (temp_ret);
+	*input = str_to_env_entry(keyval);
+	if (input->key == NULL)
+		return (MALLOC_ERROR);
 	return (0);
 }
