+++ sources/minishell.c
@@ -6,13 +6,13 @@
 /*   By: tchoquet <tchoquet@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/06 22:20:59 by tchoquet          #+#    #+#             */
-/*   Updated: 2023/08/21 11:39:10 by tchoquet         ###   ########.fr       */
+/*   Updated: 2023/08/23 18:49:18 by tchoquet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-#include "environment.h"
 #include "parser.h"
+#include "environment.h"
 
 void	minishell_loop(void)
 {
@@ -40,4 +40,27 @@ void	minishell_loop(void)
 		}
 		free((void *)cmd);
 	}
+}
+
+t_ast	*parse_cmd(const char *cmd)
+{
+	t_ast		*ast;
+	t_toklist	*token_list;
+
+	token_list = make_toklist(cmd);
+	if (token_list != NULL)
+	{
+		ast = make_ast(token_list);
+		if (ast != NULL)
+		{
+			ft_lstclear((t_list **)&token_list, NULL);
+			return (ast);
+		}
+		else
+		{
+			set_last_error(print_error(MALLOC_ERROR));
+			ft_lstclear((t_list **)token_list, &free_token);
+		}
+	}
+	return (NULL);
 }
