+++ sources/parser.c
@@ -6,13 +6,11 @@
 /*   By: tchoquet <tchoquet@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/02 20:24:15 by tchoquet          #+#    #+#             */
-/*   Updated: 2023/08/21 12:26:04 by tchoquet         ###   ########.fr       */
+/*   Updated: 2023/08/23 18:28:22 by tchoquet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parser.h"
-#include "environment.h"
-#include "expander.h"
 
 int		add_cmd_back(t_token *token, t_ast *tree);
 int		add_io_back(t_toklist **token_lst, t_ast *tree);
@@ -46,29 +44,4 @@ t_ast	*make_ast(t_toklist *toklist)
 void	clean_ast(t_ast *ast)
 {
 	btr_clear(btr_get_root((t_btree *)ast), &free_token);
-}
-
-t_ast	*parse_cmd(const char *cmd)
-{
-	t_ast		*ast;
-	t_toklist	*token_list;
-
-	token_list = make_toklist(cmd);
-	if (token_list != NULL)
-	{
-		ast = make_ast(token_list);
-		if (ast != NULL)
-		{
-			ft_lstclear((t_list **)&token_list, NULL);
-			if (expand_ast(ast) == 0)
-				return (ast);
-			clean_ast(ast);
-		}
-		else
-		{
-			set_last_error(print_error(MALLOC_ERROR));
-			clean_toklist(&token_list);
-		}
-	}
-	return (NULL);
 }
